Fix part of #59: Update Dagger structure & version from 2.28.1 to 2.41 (#4931)
## Explanation
Fix part of #59

This PR introduces a simplification to Dagger setup throughout the Bazel
build graph by defining a single top-level ``dagger_rules()`` invocation
rather than introducing up to one per directory. The new approach
actually matches Dagger documentation for how it should be setup:
https://github.com/google/dagger/blob/f41033cc448eb7bdb83af2356c8802f1208d1824/examples/bazel/BUILD#L20.

Semantically, this mainly serves to centralize the exported Dagger
dependencies to a single library. Depending on this new top-level
``//:dagger`` library will still enable Dagger annotation processing and
code generation for the library depending on it, so it's not expected to
result in any significant performance improvements for the build graph.

Note that this PR also makes some attempt at removing cases when Dagger
isn't needed as a dependency at all (though this will be done in much
more significant detail in downstream PRs), and in some cases Dagger was
added just for implicitly depending on javax annotations (which can be
depended on directly).

There are also a number of version upgrades happening in this PR in
preparation for upgrading to Kotlin 1.6 (in #4937) since older Dagger
has some compatibility issues with newer Kotlin:
- Dagger is being upgraded to 2.41 from the current 2.28.1.
- The implicit Kotlin stdlib-jdk8 dependency is being bumped from 1.4.10
to 1.5.32 (this could be significant, but note that we are already using
the 1.5.0 base Kotlin SDK). Note that this resulted in some additional
Proguard exemptions for production builds. These *should* be okay to
ignore, but they will be reevaluated in #4937.
- Guava is being specially handled to force the Android version of Guava
(and specifically 31.0.1 which is needed by the newer version of
Dagger). Without this override the JRE version would be used which would
cause Desugaring and runtime problems in production builds.
- Other dependencies were updated or explicitly listed, as needed, for
the Dagger & Guava changes:
  - com.google.errorprone:error_prone_annotations: 2.9.0 -> 2.11.0
  - com.google.guava:failureaccess: 1.0.1 (currently used version)
  - com.google.j2objc:j2objc-annotations: 1.3 (currently used version)
- org.checkerframework:checker-compat-qual: 2.5.5 (currently used
version)
  - org.checkerframework:checker-qual: 3.13.0 -> 3.21.3

It was noticed that the tracking for Guava's third-party license was
removed. This is due to Guava now being explicitly depended on rather
than being managed by rules_jvm_external. #5397 was filed to track
properly including all third-party licenses, not just those from Maven.
Addressing that issue will result in Guava's license being included
again in production builds.

Finally, a ``ExplorationProgressListener`` binding was removed from
``ExplorationProgressModuleTest`` because it was noticed during
development that it isn't actually used.

## Essential Checklist
- [x] The PR title and explanation each start with "Fix #bugnum: " (If
this PR fixes part of an issue, prefix the title with "Fix part of
#bugnum: ...".)
- [x] Any changes to
[scripts/assets](https://github.com/oppia/oppia-android/tree/develop/scripts/assets)
files have their rationale included in the PR explanation.
- [x] The PR follows the [style
guide](https://github.com/oppia/oppia-android/wiki/Coding-style-guide).
- [x] The PR does not contain any unnecessary code changes from Android
Studio
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#undo-unnecessary-changes)).
- [x] The PR is made from a branch that's **not** called "develop" and
is up-to-date with "develop".
- [x] The PR is **assigned** to the appropriate reviewers
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#clarification-regarding-assignees-and-reviewers-section)).

## For UI-specific PRs only
N/A -- build infrastructure-only change.

---------

Co-authored-by: Adhiambo Peres <59600948+adhiamboperes@users.noreply.github.com>
Co-authored-by: Sean Lip <sean@seanlip.org>