Fix #5428: Remove kotlin-android-extensions plugin (#5429)
## Explanation
Fixes #5428

This PR removes all occurrences of the kotlin-android-extensions plugin.

See the issue for specifics, but it seems the plugin wasn't actually
being used anywhere (due to it never being enabled in Bazel and the old
code that depended on it being removed years ago), so this only affects
Gradle builds & tests. It should be a mostly benign change.

## Essential Checklist
- [x] The PR title and explanation each start with "Fix #bugnum: " (If
this PR fixes part of an issue, prefix the title with "Fix part of
#bugnum: ...".)
- [x] Any changes to
[scripts/assets](https://github.com/oppia/oppia-android/tree/develop/scripts/assets)
files have their rationale included in the PR explanation.
- [x] The PR follows the [style
guide](https://github.com/oppia/oppia-android/wiki/Coding-style-guide).
- [x] The PR does not contain any unnecessary code changes from Android
Studio
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#undo-unnecessary-changes)).
- [x] The PR is made from a branch that's **not** called "develop" and
is up-to-date with "develop".
- [x] The PR is **assigned** to the appropriate reviewers
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#clarification-regarding-assignees-and-reviewers-section)).

## For UI-specific PRs only
This isn't expected to impact the user experience at all. It's possible
the final build changes a bit, but that should be innocuous since the
plugin isn't actually being used for anything.