Add support for classrooms.
Note that this is slightly out-of-date with develop, so a follow-up
consolidation will be needed.

Additionally, this also makes effort toward allowing required language
enforcement to be re-enabled, but much more work is still needed here.
Certain content IDs have been silenced for failures due to
incompatibilities that will need to be fixed either in content or in the
import pipeline.

This also introduces support for caching latest versions of structures
which will help some with repeat performance (but it simplified other
aspects of topic pack construction).

Finally, a bunch of thoughts were added to DownloadLessons via a block
comment for future changes that should be considered as the download
script is finalized (to improve the script's robustness and
maintainability as lesson structures and requirements continue to change
over time).