Replicating the unit test yaml implementation for code coverages
Note: This is a blind copy paste of the unit test implementation and this needs a lot of refining to get review ready. To self: 1. The coverages should only run after all tests have passed 2. Not sure if caching would be needed here or its necessary to understand caching here 3. Same with Oppia Build Tests, since those will require targets not sure if we will need to have an implementation for again mapping them to tests or can we run coverages directly in run coverage job, 4. The output files could possibly be moved here as paths or make sure to use the same output path here (better to move here to not cause any confusion) 5. The final .md report and the status check results should be reflected here to properly set pass and fail cases and to handle uploading of comments