Increasing the min threshold to 90 to see if the Coverage Check fails
Additionally it is noted that the files that do not have test files were added to the anomaly cases, I think it would be more appropriate to have the coverage check status set to false over there and notify the developers since there was a file that doesn't have test the coverage fails alert. Though we assume that script check checks these cases and proceed it would be better to have it as a case within the script itself to make it not dependant on other scripts or checks.