Replacing the bazel coverage command with bazel test --collect coverage to make use of the same size and shard count values for coverages
Testing it with MathModelTest does run coverage in shards while adding large and shard counts, but they do produce multiple shard coverage.dat files which on viewing seem to have entire coverage data but I just want to make sure that they are not produced as separate results, if they do then that will be a mess and we will need to figure out a way to combine all those data (would be better if bazel has that functionality (I think they do). Also testing with RunCoverageTests is taking over half an hour  with sharding so ci should be a better place to view those results.