Roughly organized layout / implementation of the flow of the protos and report generation
The base flow is RunCoverage runs CoverageRunner on each file and acquires a single or a list of protos based on the relationship with targets and test files. With the collected protos, it can either have a success (details) or a failure case. If any of them is a failure then it returns a failure for the file's test target, if the file has a success case then aggregation occurs to account incase of multi relationship cases. Till this each file will now have their own 1 single coverage report. And each of them are finally combined as one single container of coverage. This container is then passed to the CoverageReporter. Here the coverage report container is taken as one single unit and looped for each coverage report for both html and md. With html for each success case a html file needs to be created. With md report both success and failure cases should be taken and included in the final report as just one single report in one common spot (this is included just to maintain the standard flow other than that and if needed debugging, the md woudln't serve anything useful for local dev works)