Updated tests with updated md template and fixed few review comments
Needs a lot of iterations with test cases and also with the review comment to have specific testing scenerios, that should be the way but would require re iterating again. And also have calls for test content cases to simply it. Also the report structure tests too are not completely tested locally will update with ci reports as they should be much faster