Address self-review comments.
This removes code that was moved to a downstream PR, adds missing test
coverage for the new exemption behavior, and improves the messaging for
the new exemption behavior.

Also, updates the shard count for RunCoverageTest since 4 seemed to be a
bit too few when running the test locally since the tests can take quite
some time to complete.