Mapping tests to targets right in the compute changed files
This catches few things 1. Changed the source list returned earlier in the list of files pr to just return a string, earlier tried having .single() call on them as that could confirm it returning just one string path from the list, but turns out there are files that are not properly placed in the source code causing the mapping not appropriate so throwing an error would halt it, instead for those files a null is thrown and then filtered out. 2. Used distinct to just have one single source entry passed for the run coverage so no duplicate expensive runs 3. The mapping of test to source is now also a part of the compute changed files to make sure we only send sources to the run coverage and also having done this here would help with retrieving the appropriate test targets in the retrieve test targets for later builds. All these need to be tested with test cases and verified in the ci run