Use TranslationController to back audio language.
This updates ProfileManagementController to use TranslationController as
the actual source of truth for a profile's audio language setting. It
doesn't move any code away from AudioLanguage, but it does make
OppiaLanguage the true proto backing this setting now (just in a way
hidden from UI code). Long-term, AudioLanguage should be removed.

This updates calling code that needs the audio language property to use
a new getter in ProfileManagementController. The old audioLanguage
property in Profile has been removed (which will result in a setting
regression for users, but this is considered fine since we're in beta
and it's a small regression most users are unlikely to even notice).

French and Chinese have been removed entirely from AudioLanguage since
they aren't valid languages for the app at the moment (per the
OppiaLanguage enum and supported languages configurations).

TranslationController was updated to persist written translation and
audio language settings since, before, only the app language selection
was persisted.